a6

+++ src/Kambaz/Courses/Modules/index.tsx
@@ -1,7 +1,11 @@
 import { ListGroup } from "react-bootstrap";
-import { setModules,addModule, editModule,
-  updateModule, deleteModule }
-from "./reducer";
+import {
+  setModules,
+  addModule,
+  editModule,
+  updateModule,
+  deleteModule,
+} from "./reducer";
 import { useSelector, useDispatch } from "react-redux";
 import { useEffect } from "react";
 import { BsGripVertical } from "react-icons/bs";
@@ -19,28 +23,47 @@ export default function Modules() {
   // const _modules = db.modules.filter((m: any) => m.course === cid);
   // const [modules, setModules] = useState<any>(_modules);
   const [moduleName, setModuleName] = useState("");
-  const { modules } = useSelector(
-    (state: any) => state.modulesReducer);
+  const { modules } = useSelector((state: any) => state.modulesReducer);
+  const { currentUser } = useSelector((state: any) => state.accountReducer);
   const dispatch = useDispatch();
-  const createModuleForCourse = async () => {
-    if (!cid) return;
-    const newModule = { name: moduleName, course: cid };
-    const module = await coursesClient.createModuleForCourse(cid, newModule);
-    dispatch(addModule(module));
+  
+  // Check if user is faculty or admin
+  const isFacultyOrAdmin = currentUser?.role === "FACULTY" || currentUser?.role === "ADMIN";
+  
+  const fetchModules = async () => {
+    try {
+      const modules = await coursesClient.findModulesForCourse(cid!);
+      dispatch(setModules(modules));
+    } catch (error) {
+      console.error("Error fetching modules:", error);
+    }
   };
-  const removeModule = async (moduleId: string) => {
+  useEffect(() => {
+    fetchModules();
+  }, [cid]);
+  const deleteModuleHandler = async (moduleId: string) => {
+    // Only allow faculty/admin to delete modules
+    if (!isFacultyOrAdmin) return;
+    
     await modulesClient.deleteModule(moduleId);
     dispatch(deleteModule(moduleId));
   };
-
-  const fetchModules = async () => {
-    const modules = await coursesClient.findModulesForCourse(cid as string);
-    dispatch(setModules(modules));
+ 
+  const addModuleHandler = async () => {
+    // Only allow faculty/admin to add modules
+    if (!isFacultyOrAdmin) return;
+    
+    const newModule = await coursesClient.createModuleForCourse(cid!, {
+      name: moduleName,
+      course: cid,
+    });
+    dispatch(addModule(newModule));
+    setModuleName("");
   };
-
-
-
-  const saveModule = async (module: any) => {
+  const updateModuleHandler = async (module: any) => {
+    // Only allow faculty/admin to update modules
+    if (!isFacultyOrAdmin) return;
+    
     await modulesClient.updateModule(module);
     dispatch(updateModule(module));
   };
@@ -78,34 +101,47 @@ export default function Modules() {
 
   return (
     <div className="container-fluid">
-      <div>
-      <ModulesControls setModuleName={setModuleName} moduleName={moduleName} addModule={createModuleForCourse} />
-
-      </div>
+      {/* Only show module controls for faculty/admin */}
+      {isFacultyOrAdmin && (
+        <div>
+          <ModulesControls
+            addModule={addModuleHandler}
+            setModuleName={setModuleName}
+            moduleName={moduleName}
+          />
+        </div>
+      )}
 
       <ListGroup className="rounded-0 container-fluid">
-        {modules.map((module : any) => (
+        {modules.map((module: any) => (
           <ListGroup.Item
             className="wd-module
                       p-0 mb-5 fs-5 border-gray"
           >
             <div className="wd-title p-3 ps-2 bg-secondary">
-            {!module.editing && module.name}
-      { module.editing && (
-        <input className="form-control w-50 d-inline-block"
-          onChange={(e) => dispatch(
-            updateModule({ ...module, name: e.target.value }))}
-          onKeyDown={(e) => {
-            if (e.key === "Enter") {
-              saveModule({ ...module, editing: false });
-
-            }}}
-          defaultValue={module.name} />)}
-              <ModuleControlButtons moduleId={module._id}
-        deleteModule={(moduleId) => removeModule(moduleId)}
-
-        editModule={(moduleId) =>
-          dispatch(editModule(moduleId))} />
+              {!module.editing && module.name}
+              {module.editing && isFacultyOrAdmin && (
+                <input
+                  className="form-control w-50 d-inline-block"
+                  onChange={(e) =>
+                    updateModuleHandler({ ...module, name: e.target.value })
+                  }
+                  onKeyDown={(e) => {
+                    if (e.key === "Enter") {
+                      updateModuleHandler({ ...module, editing: false });
+                    }
+                  }}
+                  defaultValue={module.name}
+                />
+              )}
+              {/* Only show control buttons for faculty/admin */}
+              {isFacultyOrAdmin && (
+                <ModuleControlButtons
+                  moduleId={module._id}
+                  deleteModule={(moduleId) => deleteModuleHandler(moduleId)}
+                  editModule={(moduleId) => dispatch(editModule(moduleId))}
+                />
+              )}
             </div>
 
             {module.lessons && (
@@ -114,7 +150,8 @@ export default function Modules() {
                   <ListGroup.Item className="wd-lesson p-3 ps-1">
                     <BsGripVertical className="me-2 fs-3" />
                     {lesson.name}
-                    <LessonControlButtons />
+                    {/* Only show lesson control buttons for faculty/admin */}
+                    {isFacultyOrAdmin && <LessonControlButtons />}
                   </ListGroup.Item>
                 ))}
               </ListGroup>
@@ -122,7 +159,6 @@ export default function Modules() {
           </ListGroup.Item>
         ))}
       </ListGroup>
-      
     </div>
   );
 }
